enfin fini

+++ controllers/controllerssauce.js
@@ -107,23 +107,33 @@ exports.likeAndDislike = (req, res, next) => {
 exports.modifySauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
+      const oldUrl = sauce.imageUrl;
       const filename = sauce.imageUrl.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
-        const sauceObject = req.file
-          ? {
-              ...JSON.parse(req.body.sauce),
-              imageUrl: `${req.protocol}://${req.get("host")}/images/${
-                req.file.filename
-              }`,
-            }
-          : { ...req.body };
+      if (req.file) {
+        fs.unlink(`images/${filename}`, () => {
+          const sauceObject = {
+            ...JSON.parse(req.body.sauce),
+            imageUrl: `${req.protocol}://${req.get("host")}/images/${
+              req.file.filename
+            }`,
+          };
+          Sauce.updateOne(
+            { _id: req.params.id },
+            { ...sauceObject, _id: req.params.id }
+          )
+            .then(() => res.status(200).json({ message: "Sauce mise à jour!" }))
+            .catch((error) => res.status(400).json({ error }));
+        });
+      } else {
+        const newItem = req.body;
+        newItem.imageUrl = oldUrl;
         Sauce.updateOne(
           { _id: req.params.id },
-          { ...sauceObject, _id: req.params.id }
+          { ...newItem, imageUrl: oldUrl, _id: req.params.id }
         )
           .then(() => res.status(200).json({ message: "Sauce mise à jour!" }))
           .catch((error) => res.status(400).json({ error }));
-      });
+      }
     })
     .catch((error) => res.status(500).json({ error }));
 };
